POSIX says timeval's have two long's
SCCS-vsn: 8.5

+++ sys/sys/time.h
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)time.h	8.4 (Berkeley) 03/20/95
+ *	@(#)time.h	8.5 (Berkeley) 05/04/95
  */
 
 #ifndef _SYS_TIME_H_
@@ -43,8 +43,8 @@
  * and used in other calls.
  */
 struct timeval {
-	int32_t	tv_sec;		/* seconds */
-	int32_t	tv_usec;	/* and microseconds */
+	long	tv_sec;		/* seconds */
+	long	tv_usec;	/* and microseconds */
 };
 
 /*
